+++ CHANGELOG.md
@@ -5,13 +5,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- New method `Toolchain::rustup_binary` to allow running arbitrary binaries managed by rustup. Before, only `rustc` and `cargo` could be run.
+
 ## [0.12.0] - 2021-01-28
 
 ### Added
 
 - New variant `PrepareError::MissingDependencies`, returned during the prepare
   step when a dependency does not exist.
-
 ### Changed
 
 - Path dependencies are no longer removed from `Cargo.toml` during the prepare